feat: atualiza projeto feedeliza

+++ migrations/20250811223000-create-stocks-table.js
@@ -41,6 +41,8 @@ module.exports = {
   },
 
   down: async (queryInterface, Sequelize) => {
-    await queryInterface.dropTable('stocks');
-  }
+    const tableExists = await queryInterface.tableExists('stocks');
+    if (tableExists) {
+      await queryInterface.dropTable('stocks');
+    }
 };
